Update MPR/HTTP deps - Fix etag creationg - Fix uninitialized allYielded.

+++ src/deps/http/httpLib.c
@@ -8612,7 +8612,7 @@ void httpMapFile(HttpConn *conn, HttpRoute *route)
     info = &tx->fileInfo;
     mprGetPathInfo(tx->filename, info);
     if (info->valid) {
-        tx->etag = sfmt("\"%Lx-%Lx-%Lx\"", info->inode, info->size, info->mtime);
+        tx->etag = sfmt("\"%Lx-%Lx-%Lx\"", (int64) info->inode, (int64) info->size, (int64) info->mtime);
     }
     LOG(7, "mapFile uri \"%s\", filename: \"%s\", extension: \"%s\"", rx->uri, tx->filename, tx->ext);
 }
